added redis hotqueue output for scraped data
will have another script that runs with the webapp to pull items off the redis queue and stuff them into mysql.
doing this with redis so that we can easily scale if necessary, and to prevent mysql from blocking the scraper from moving forward.

This also allows us to potentially feed data in from other sources not just BDD.